+++ src/main/java/com/be/sportizebe/domain/user/controller/UserController.java
@@ -2,7 +2,7 @@ package com.be.sportizebe.domain.user.controller;
 
 import com.be.sportizebe.domain.user.dto.request.SignUpRequest;
 import com.be.sportizebe.domain.user.dto.response.SignUpResponse;
-import com.be.sportizebe.domain.user.service.UserService;
+import com.be.sportizebe.domain.user.service.UserServiceImpl;
 import com.be.sportizebe.global.response.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -21,7 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "user", description = "사용자 관련 API")
 public class UserController {
 
-    private final UserService userService;
+    private final UserServiceImpl userService;
 
     @PostMapping("/signup")
     @Operation(summary = "회원가입", description = "이메일과 비밀번호로 회원가입")
